컴파일 오류난 파일들 수정

+++ MovieManager.cpp
@@ -10,7 +10,7 @@ void MovieManager::addMovie(const Movie& movie) {
 }
 
 Movie* MovieManager::findByTitle(const std::string& title) {
-    for (int i=0; i<movies.size(); i++) {
+    for (size_t i=0; i<movies.size(); i++) {
         if (movies[i].getTitle() == title) {
             return &movies[i];
         }
@@ -28,7 +28,7 @@ void MovieManager::printAll() const {
         return;
     }
 
-    for (int i=0; i<movies.size(); i++) {
+    for (size_t i=0; i<movies.size(); i++) {
         movies[i].display();
     }
 }
